Move test_helpers to the sources pkg.
